requirements.txt
gradio>=4.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
python-dotenv>=1.0.0